+++ src/main/java/com/ztdx/eams/query/SystemQuery.java
@@ -326,7 +326,7 @@ public class SystemQuery {
                 sysFonds.ORDER_NUMBER.as("orderNumber"),
                 sysFonds.REMARK.as("remark"))
                 .from(sysFonds)
-                .where(sysFonds.GMT_DELETED.equal((byte) 0))
+                .where(sysFonds.GMT_DELETED.equal(0))
                 .orderBy(sysFonds.ORDER_NUMBER)
                 .fetch().intoMaps();
         return dataList;
